requirements.txt
python-mnist==0.5
pandas==0.22.0
sklearn==0.0
scikit-learn==0.19.2
scipy==1.1.0
matplotlib==2.2.2